% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aln}
\alias{aln}
\title{Klebiella pneumoniae ST258 isolate variant alignment}
\format{
A DNAbin object with 413 rows and 83976 nucelotide sites.
}
\usage{
aln
}
\description{
Variant alignment for Klebsiella pneumoniae ST258 isolates in the study.
Isolates were collected from long-term acute care hospitals across the US.
Corresponding manuscript: https://aac.asm.org/content/63/11/e01622-19
}
\keyword{datasets}
