% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{days_between_isolates}
\alias{days_between_isolates}
\title{Difference in days between isolate collection dates}
\format{
A data frame with 85078 rows and 3 variables.
}
\usage{
days_between_isolates
}
\description{
For all pairs of isolates
Corresponding manuscript: https://aac.asm.org/content/63/11/e01622-19
}
\keyword{datasets}
