% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dists}
\alias{dists}
\title{Klebiella pneumoniae ST258 isolate pairwise SNV distance}
\format{
A data frame with 413 rows and 413 columns.
}
\usage{
dists
}
\description{
Pairwise single nucleotide variant (SNV) distance matrix for
Klebsiella pneumoniae ST258 isolates in the study.
samples were collected from long-term acute care hospitals across the US.
Corresponding manuscript: https://aac.asm.org/content/63/11/e01622-19
}
\keyword{datasets}
