% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dists_mini}
\alias{dists_mini}
\title{Subsetted pairwise SNV distances}
\format{
A data frame.
}
\usage{
dists_mini
}
\description{
Subset of \code{dists}. Small version for examples.
}
\keyword{datasets}
