% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facil_coord}
\alias{facil_coord}
\title{Deidentified facility longitude and latitude}
\format{
A data frame.
}
\usage{
facil_coord
}
\description{
Deidentified facility longitude and latitude for plotting
}
\keyword{datasets}
