% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsp}
\alias{fsp}
\title{Klebiella pneumoniae ST258 gene flow}
\format{
A data frame with 413 rows and 413 columns.
}
\usage{
fsp
}
\description{
Gene flow (fsp) between facilities for Klebsiella pneumoniae ST258 isolates.
samples were collected from long-term acute care hospitals across the US.
Corresponding manuscript: https://aac.asm.org/content/63/11/e01622-19
}
\keyword{datasets}
