% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect_flow.R
\name{get_indirect_flow}
\alias{get_indirect_flow}
\title{Calculate indirect patient flow from patient transfer network}
\usage{
get_indirect_flow(pt_trans_net)
}
\arguments{
\item{pt_trans_net}{a dataframe representing a patient transfer network of 3 cols: 'source_facil', 'dest_facil, and 'n_transfers' (code doesn't support missing paths, any missing paths will be represented by 0s)}
}
\value{
facility x facility matrix of metric of patient flow between each facility pair
}
\description{
Calculate indirect patient flow from patient transfer network
}
\examples{
get_indirect_flow(pt_trans_net)
}
