% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facility_fsp.R
\name{get_long_form}
\alias{get_long_form}
\title{Make symmetric matrix long-form where each row represents a cell in the matrix}
\usage{
get_long_form(facil_dist)
}
\arguments{
\item{facil_dist}{symmetric matrix that you want to convert to long form}
}
\value{
long form data where each row represents a cell in the matrix (rowname, column name, value)
}
\description{
Make symmetric matrix long-form where each row represents a cell in the matrix
}
\examples{
get_long_form(dists)
get_long_form(fsp)
}
