% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patient_flow.R
\name{get_patient_flow}
\alias{get_patient_flow}
\title{Summarize inter-facility patient transfer network}
\usage{
get_patient_flow(edge_df, paths = FALSE)
}
\arguments{
\item{edge_df}{a dataframe representing a patient transfer network of 3 cols: 'source_facil', 'dest_facil, and 'n_transfers' (code doesn't support missing paths, any missing paths will be represented by 0s)}

\item{paths}{boolean value, TRUE if you want the shortest paths returned, FALSE if you don't}
}
\value{
the number of direct patient transfers and indirect flow metrics between each facility pair. If paths = TRUE, a list of summary (pt_trans_summary) and shortest paths used (paths).
}
\description{
Summarize inter-facility patient transfer network
}
\examples{
get_patient_flow(edge_df = pt_trans_df)
}
