% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patient_transfers.R
\name{get_patient_transfers}
\alias{get_patient_transfers}
\title{Get summary of patient transfer network and closely-related isolates}
\usage{
get_patient_transfers(
  pt_trans_net,
  snv_dists = NULL,
  dists = NULL,
  locs = NULL,
  pt = NULL,
  thresh = 10,
  paths = FALSE
)
}
\arguments{
\item{pt_trans_net}{a dataframe representing a patient transfer network of 3 cols: 'source_facil', 'dest_facil, and 'n_transfers' (code doesn't support missing paths, any missing paths will be represented by 0s)}

\item{snv_dists}{the output object of the get_snv_dists function}

\item{dists}{a SNV distance matrix returned by the dist.dna function from the ape package}

\item{locs}{a named vector of locations of isolates (e.g. facility of isolation), with the name being the sample ID}

\item{pt}{a named vector of patient that isolate was taken from with the name being sample ID (optional)}

\item{thresh}{SNV thresholds to use}

\item{paths}{boolean value, TRUE if you want the shortest paths returned, FALSE if you don't}
}
\value{
a summary of number of closely related isolate pairs and number of direct patient transfers and indirect flow metrics between each facility pair. If paths = TRUE, a list of summary (pt_trans_summary) and shortest paths used (paths).
}
\description{
Get summary of patient transfer network and closely-related isolates
}
\examples{
\dontrun{
locs <- metadata \%>\% dplyr::select(isolate_id, facility) \%>\% tibble::deframe()
get_patient_transfers(pt_trans_net = pt_trans_df, dists = dists, locs = locs)
}
}
