% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snv_dists.R
\name{get_snv_dists}
\alias{get_snv_dists}
\title{Get data frame of inter- and intra-facility pariwise SNV distances}
\usage{
get_snv_dists(dists, locs)
}
\arguments{
\item{dists}{a SNV distance matrix returned by the dist.dna function from the ape package}

\item{locs}{a named vector of locations of isolates (e.g. facility of isolation), with the name being the sample ID}
}
\value{
a data.frame of isolate pairs, their SNV distance, and labeled as either inter- or intra-facility pairs.
}
\description{
Get data frame of inter- and intra-facility pariwise SNV distances
}
\examples{
\dontrun{
locs <- metadata \%>\% dplyr::select(isolate_id, facility) \%>\% tibble::deframe()
snv_dists <- get_snv_dists(dists, locs)
}
}
