% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facility_fsp.R
\name{long_form}
\alias{long_form}
\title{Make symmetric matrix long-form where each row represents a cell in the matrix}
\usage{
long_form(facil_dist)
}
\arguments{
\item{facil_dist}{}
}
\value{
long form data where each row represents a cell in the matrix
}
\description{
Make symmetric matrix long-form where each row represents a cell in the matrix
}
\examples{
long_form(facil_dist)
}
