% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_long_form}
\alias{make_long_form}
\title{Make matrix long-form where each row represents a cell in the matrix}
\usage{
make_long_form(facil_dist, col_names = c("loc1", "loc2", "fsp"))
}
\arguments{
\item{facil_dist}{symmetric or asymmetric matrix that you want to convert to long form.}

\item{col_names}{Column names for the output data frame (3 columns - rownames, colnames, values)}
}
\value{
long form data where each row represents a cell in the matrix (rowname, column name, value)
}
\description{
Each row represents a cell in the matrix (rowname, column name, value).
If the matrix is symmetric, there will be one row for each symmetric pair (i.e. the number of rows in the long-form data frame will be half the number of elements in the input matrix).
If the matrix is asymmetric, all rows will be kept (i.e. the number of rows in the long-form data frame will be equal to the number of elements in the input matrix).
}
\examples{
make_long_form(fsp)
}
