% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Klebiella pneumoniae ST258 isolate metadata}
\format{
A data frame with 413 rows and 4 variables.
}
\usage{
metadata
}
\description{
Metadata including isoalte ID, patient ID, facility, and isolate collection
date of Klebsiella pneumoniae ST258 isolates collected from long-term acute
care hospitals across the US.
Corresponding manuscript: https://aac.asm.org/content/63/11/e01622-19
}
\keyword{datasets}
