% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_mini}
\alias{metadata_mini}
\title{Subsetted metadata}
\format{
A data frame

A data frame.
}
\usage{
metadata_mini

metadata_mini
}
\description{
Small version for examples.

Subset of \code{metadata}. Small version for examples.
}
\keyword{datasets}
