% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_transfer.R
\name{patient_transfer}
\alias{patient_transfer}
\title{Get summary of patient transfer network and closely-related isolates}
\usage{
patient_transfer(
  pt_trans_net,
  snv_dists = NULL,
  dists = NULL,
  locs = NULL,
  pt = NULL,
  thresh = 10
)
}
\arguments{
\item{pt_trans_net}{a dataframe representing a patient transfer network of 3 cols: 'source_facil', 'dest_facil, and 'n_transfers'}

\item{snv_dists}{the output object of the get_snv_dists function}

\item{dists}{a SNV distance matrix returned by the dist.dna function from the ape package}

\item{locs}{a named vector of locations of isolates (e.g. facility of isolation), with the name being the sample ID}

\item{pt}{a named vector of patient that isolate was taken from with the name being sample ID (optional)}

\item{threshs}{SNV thresholds to use}
}
\value{
fraction of intra-facility pairs for different snv thresholds, lowest threshold represents lowest snv_dist in your data
}
\description{
Get summary of patient transfer network and closely-related isolates
}
\examples{
a patient transfer network and either input a snv_dists object that is the output of the get_snv_dists function or input a SNV distance matrix (made by ape::dists.dna) and a named vector of isolate locations and optionally isolate patient IDs.
}
