% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_flow}
\alias{pt_flow}
\title{Patient flow between facilities}
\format{
A data frame with 11 rows and 11 columns.
}
\usage{
pt_flow
}
\description{
Patient flow between California facilities calculated from a patient
transfer network of the number of patients transferred between
facilities over the course of a year.
Corresponding manuscript: https://aac.asm.org/content/63/11/e01622-19
}
\keyword{datasets}
