% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_trans_df}
\alias{pt_trans_df}
\title{Made up patient transfer dataframe between facilities}
\format{
A data frame with 11 rows and 11 columns.
}
\usage{
pt_trans_df
}
\description{
Made up patient transfers between facilities based on the true patient
flow network. We can't provide the original data here.
Corresponding manuscript: https://aac.asm.org/content/63/11/e01622-19
}
\keyword{datasets}
