% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reverse_list_str}
\alias{reverse_list_str}
\title{Reverse list structure}
\usage{
reverse_list_str(ls)
}
\arguments{
\item{ls}{list you want to reverse}
}
\value{
reversed list
}
\description{
Reverse list structure
}
\details{
Reference with example: https://stackoverflow.com/questions/15263146/revert-list-structure
}
