% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_pairs.R
\name{subset_pairs}
\alias{subset_pairs}
\title{subset_pairs}
\usage{
subset_pairs(snv_dists)
}
\arguments{
\item{snv_dists}{the output object of the get_snv_dists function}
}
\value{
a data.frame of isolate pairs subsetted to one row representing each pair
}
\description{
subset_pairs
}
\examples{
subset_pairs(snv_dists)
}
