% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tr}
\alias{tr}
\title{Klebiella pneumoniae ST258 isolate phylogenetic tree}
\format{
A phylo object with 413 tips.
}
\usage{
tr
}
\description{
Phylogenetic tree of Klebsiella pneumoniae ST258 isolates in the study.
Isolates were collected from long-term acute care hospitals across the US.
Corresponding manuscript: https://aac.asm.org/content/63/11/e01622-19
}
\keyword{datasets}
