% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facility_fsp.R
\name{get_facility_fsp}
\alias{get_facility_fsp}
\title{Make facility x facility matrix with Fsp values. Calculates fst as described in Donker et al. 2017}
\usage{
get_facility_fsp(fasta, locs, form = "matrix")
}
\arguments{
\item{fasta}{ape DNAbin object (i.e. from fasta file of SNPs) using read.fasta}

\item{locs}{locations names for pairwise comparison}

\item{form}{can be "long" or "matrix", default is matrix. Determines output format}
}
\value{
matrix of facility x facility matrix with Fsp values. Only bi-allelic sites. Fsp values bween 0 (HP=HS) and 1 (Hp = 0)
}
\description{
Make facility x facility matrix with Fsp values. Calculates fst as described in Donker et al. 2017
}
\examples{
\dontrun{get_facility_fsp(fasta, locs, form = "matrix")}
}
