% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frac_intra.R
\name{get_frac_intra}
\alias{get_frac_intra}
\title{Get fraction of intra-facility pairs for different snv thresholds}
\usage{
get_frac_intra(snv_dists)
}
\arguments{
\item{snv_dists}{the output object of the get_snv_dists function}
}
\value{
fraction of intra-facility pairs for different snv thresholds, lowest threshold represents lowest snv_dist in your data
}
\description{
Get fraction of intra-facility pairs for different snv thresholds
}
\examples{
\dontrun{
locs <- metadata \%>\% dplyr::select(isolate_id, facility) \%>\% tibble::deframe()
get_frac_intra(dists = dists, locs = locs)
}
}
