% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frac_intra.R
\name{get_frac_intra}
\alias{get_frac_intra}
\title{Get fraction of intra-facility pairs for different snv thresholds}
\usage{
get_frac_intra(
  snv_dists = NULL,
  dists = NULL,
  locs = NULL,
  pt = NULL,
  pt_trans_net = NULL
)
}
\arguments{
\item{snv_dists}{the output object of the get_snv_dists function}

\item{dists}{a SNV distance matrix returned by the dist.dna function from the ape package}

\item{locs}{a named vector of locations of isolates (e.g. facility of isolation), with the name being the sample ID}

\item{pt}{a named vector of patient that isolate was taken from with the name being sample ID (optional)}

\item{pt_trans_net}{a dataframe representing a patient transfer network of 3 cols: 'source_facil', 'dest_facil, and 'n_transfers' (code doesn't support missing paths, any missing paths will be represented by 0s)}
}
\value{
fraction of intra-facility pairs for different snv thresholds, lowest threshold represents lowest snv_dist in your data
}
\description{
Get fraction of intra-facility pairs for different snv thresholds
}
\examples{
\dontrun{
locs <- metadata \%>\% dplyr::select(isolate_id, facility) \%>\% tibble::deframe()
get_frac_intra(dists = dists, locs = locs)
}
}
