% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frac_intra.R
\name{get_frac_intra}
\alias{get_frac_intra}
\title{Get fraction of intra-facility pairs for different SNV thresholds}
\usage{
get_frac_intra(snv_dists)
}
\arguments{
\item{snv_dists}{the output object of the get_snv_dists function}
}
\value{
fraction of intra-facility pairs for each SNV distance in dataset
}
\description{
Get fraction of intra-facility pairs for different SNV thresholds
}
\examples{
\dontrun{
locs <- metadata \%>\% dplyr::select(sample_id, facility) \%>\% tibble::deframe()
snv_dists <- get_snv_dists(dists, locs)
get_frac_intra(snv_dists)
}
}
