% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frac_intra.R
\name{get_frac_intra}
\alias{get_frac_intra}
\title{Get fraction of intra-facility pairs for different snv thresholds}
\usage{
get_frac_intra(
  snv_dists = NULL,
  dists = NULL,
  locs = NULL,
  pt = NULL,
  threshs = seq(1, 50, 1)
)
}
\arguments{
\item{snv_dists}{the output object of the get_snv_dists function}

\item{dists}{a SNV distance matrix returned by the dist.dna function from the ape package}

\item{locs}{a named vector of locations of isolates (e.g. facility of isolation), with the name being the sample ID}

\item{pt}{a named vector of patient that isolate was taken from with the name being sample ID (optional)}

\item{threshs}{SNV thresholds to use}
}
\value{
fraction of intra-facility pairs for different snv thresholds
}
\description{
Get fraction of intra-facility pairs for different snv thresholds
}
\examples{
either input a snv_dists object that is the output of the get_snv_dists function or input a SNV distance matrix (made by ape::dists.dna) and a named vector of isolate locations and optionally isolate patient IDs.
}
