% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snv_dists.R
\name{get_snv_dists}
\alias{get_snv_dists}
\title{Get matrix of inter- and intra-facility pariwise snv distances}
\usage{
get_snv_dists(dists, locs, pt = NULL, pt_trans_net = NULL)
}
\arguments{
\item{dists}{a SNV distance matrix returned by the dist.dna function from the ape package}

\item{locs}{a named vector of locations of isolates (e.g. facility of isolation), with the name being the sample ID}

\item{pt}{a named vector of patient that isolate was taken from with the name being sample ID (optional)}

\item{pt_trans_net}{a data.frame representing a patient transfer network of 3 cols: 'source_facil', 'dest_facil, and 'n_transfers'}
}
\value{
a data.frame of isolate pairs, their SNV distance, and labeled as either inter- or intra-facility pairs
}
\description{
Get matrix of inter- and intra-facility pariwise snv distances
}
\examples{
to reduce this dataframe to include only one row that represents each pair, use the subset_pairs() function on the output of this function
}
