% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_inter_pairs.R
\name{summarize_pairs}
\alias{summarize_pairs}
\title{Get number of closely related pairs between facilities for different SNV distance thresholds}
\usage{
summarize_pairs(snv_dists, summary_fns = c("min"), threshs = seq(5, 20, 5))
}
\arguments{
\item{snv_dists}{the output object of the get_snv_dists function}

\item{summary_fns}{vector of summary functions for pairwise distances as character strings (default: c("min"))}

\item{threshs}{SNV thresholds to use for pairwise distances (default: seq(5, 20, 5))}
}
\value{
a summary of isolate pairs between each facility pair.
}
\description{
Get number of closely related pairs between facilities for different SNV distance thresholds
}
\examples{
\dontrun{
locs <- metadata \%>\% dplyr::select(isolate_id, facility) \%>\% tibble::deframe()
snv_dists <- get_snv_dists(dists, locs)
summarize_pairs(snv_dists = snv_dists)
}
}
