% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pairs.R
\name{summarize_pairs}
\alias{summarize_pairs}
\title{Summarize extent of relatedness of intra- and inter-facility pairs by facility pair}
\usage{
summarize_pairs(pair_types, summary_fns = c("min"), threshs = seq(5, 20, 5))
}
\arguments{
\item{pair_types}{the output object of the get_pair_types function}

\item{summary_fns}{vector of summary functions for pairwise distances as character strings (default: c("min"))}

\item{threshs}{SNV thresholds to use for pairwise distances (default: seq(5, 20, 5))}
}
\value{
a summary of isolate pairs between each facility pair. leq = less than or equal to
}
\description{
Summarize extent of relatedness of intra- and inter-facility pairs by facility pair
}
\examples{
\dontrun{
locs <- metadata \%>\% dplyr::select(isolate_id, facility) \%>\% tibble::deframe()
pair_types <- get_pair_types(dists, locs)
summarize_pairs(pair_types = pair_types)
}
}
